Mohita.pdf
1. Introduction 
The To-Do List Website is a simple yet functional web-based application 
that allows users to manage daily tasks efficiently. Users can add, mark 
complete, and delete tasks. The project demonstrates the effective use of 
HTML, CSS, and JavaScript to create an interactive and visually 
appealing user experience. It is designed for beginners to understand basic 
web development concepts and for users to practice organizing tasks 
digitally.